requirements.txt
langchain>=0.3.27
langchain-openai>=0.3.31
langchain-community>=0.3.27
langchain-core>=0.3.74
streamlit==1.37.1
pyyaml==6.0.2
openai>=1.42.0
python-dotenv==1.0.1
